+++ packages/dso-toolkit/src/components/annotation-button/annotation-button.models.ts
@@ -1,9 +1,0 @@
-export interface AnnotationButton {
-  identifier: string;
-  open?: boolean;
-  dsoClick?: (e: CustomEvent<AnnotationButtonClick>) => void;
-}
-
-export interface AnnotationButtonClick {
-  originalEvent: MouseEvent;
-}