+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Zie de blogpost op https://www.dso-toolkit.nl voor meer informatie over de break
 ### Added
 * Build: Artifact integrity ([#2688](https://github.com/dso-toolkit/dso-toolkit/issues/2688))
 
+### Changed
+* Document Component: Panel implementeren ([#2765](https://github.com/dso-toolkit/dso-toolkit/issues/2765))
+
 ### Removed
 * **BREAKING** Panel: HTML/CSS implementatie verwijderen ([#1464](https://github.com/dso-toolkit/dso-toolkit/issues/1464))